New removeAttribute function.
Added a new function in DeviceModelable to let you remove an attribute.  Also fixed some minor typos.

Note that the subscript deprecated message should probably mention removeAttribute.  For now I’ve left it out, but if we do include it in the future I’d suggest renaming the function to just “remove”.